Example for Uplearn
This commit introduces a small example project for Uplearn as part of
technical application.

The exercise involves the following challenge:

Write a function fetch(url) that fetches the page corresponding to the
url and returns an object that has the following attributes:

- assets - an array of urls present in the <img> tags on the page
- links - an array of urls present in the <a> tags on the page

Assume that the code will run on a server. Assume that this work is a
part of a web app that needs to be built further by multiple development
teams and will be maintained and evolved for several years in the
future. In addition to these, make any assumptions necessary, but list
those assumptions explicitly.

The README provides a comprehensive overview of the project and its
capabilities as well as examples how to use the code.

The code was written using Elixir 1.9, however it should work with
most versions of Elixir.